refactor: relocate logic
