+++ packages/svg/src/types.ts
@@ -3,3 +3,13 @@ export interface SvgNode {
   props: Record<string, unknown>;
   children?: SvgNode[];
 }
+
+export interface XmlElement {
+  tagName: string;
+  attributes: Record<string, string>;
+  children: (XmlElement | XmlText)[];
+}
+
+export interface XmlText {
+  text: string;
+}
